+++ src/lambda/handler-runner/python-runner/PythonRunner.js
@@ -7,12 +7,13 @@ import { fileURLToPath } from 'node:url'
 import { log } from '@serverless/utils/log.js'
 
 const { parse, stringify } = JSON
-const { assign } = Object
-const { has } = Reflect
+const { assign, hasOwn } = Object
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 
 export default class PythonRunner {
+  static #payloadIdentifier = '__offline_payload__'
+
   #env = null
 
   #handlerName = null
@@ -83,9 +84,9 @@ export default class PythonRunner {
       if (
         json &&
         typeof json === 'object' &&
-        has(json, '__offline_payload__')
+        hasOwn(json, PythonRunner.#payloadIdentifier)
       ) {
-        payload = json.__offline_payload__
+        payload = json[PythonRunner.#payloadIdentifier]
         // everything else is print(), logging, ...
       } else {
         log.notice(item)