+++ tests/integration/jwt-authorizer/jwt-authorizer.test.js
@@ -6,7 +6,7 @@ import { randomBytes } from 'node:crypto'
 import { dirname, resolve } from 'node:path'
 import { fileURLToPath } from 'node:url'
 import jsonwebtoken from 'jsonwebtoken'
-import { setup, teardown } from '../_testHelpers/index.js'
+import { setup, teardown } from '../../_testHelpers/index.js'
 import { BASE_URL } from '../../config.js'
 
 const { now } = Date