+++ tests/lambda-run-mode/in-process/instances/in-process-instances.test.js
@@ -2,8 +2,8 @@ import assert from 'node:assert'
 import { dirname, resolve } from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { promisify } from 'node:util'
+import { setup, teardown } from '../../../_testHelpers/index.js'
 import { BASE_URL } from '../../../config.js'
-import { setup, teardown } from '../../../integration/_testHelpers/index.js'
 
 const setTimeoutPromise = promisify(setTimeout)