+++ tests/lambda-run-mode/worker-threads/local-env-vars/worker-threads-local-env-vars.test.js
@@ -1,15 +1,17 @@
 import assert from 'node:assert'
 import { dirname, resolve } from 'node:path'
 import { fileURLToPath } from 'node:url'
+import { setup, teardown } from '../../../_testHelpers/index.js'
 import { BASE_URL } from '../../../config.js'
-import { setup, teardown } from '../../../integration/_testHelpers/index.js'
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 
-describe('run mode with worker threads and --localEnvironmentVariables flag', function desc() {
+describe('run mode with worker threads and --localEnvironment flag', function desc() {
   beforeEach(() =>
     setup({
       env: {
+        AWS_ACCESS_KEY_ID: 'SHOULD_BE_SHARED',
+        AWS_FOOBAR: 'SHOULD_BE_SHARED',
         ENV_SHOULD_BE_SHARED: 'ENV_SHOULD_BE_SHARED',
       },
       servicePath: resolve(__dirname),
@@ -24,10 +26,13 @@ describe('run mode with worker threads and --localEnvironmentVariables flag', fu
     const response = await fetch(url)
     assert.equal(response.status, 200)
 
-    const json = await response.json()
-    assert.equal(json.env.ENV_FUNCTIONS_FOO, 'ENV_FUNCTIONS_BAR')
-    assert.equal(json.env.ENV_PROVIDER_FOO, 'ENV_PROVIDER_BAR')
-    assert.equal(json.env.ENV_SHOULD_BE_SHARED, 'ENV_SHOULD_BE_SHARED')
-    assert.equal(json.env.IS_OFFLINE, 'true')
+    const { env } = await response.json()
+
+    assert.equal(env.AWS_ACCESS_KEY_ID, 'SHOULD_BE_SHARED')
+    assert.equal(env.AWS_FOOBAR, 'SHOULD_BE_SHARED')
+    assert.equal(env.ENV_FUNCTIONS_FOO, 'ENV_FUNCTIONS_BAR')
+    assert.equal(env.ENV_PROVIDER_FOO, 'ENV_PROVIDER_BAR')
+    assert.equal(env.ENV_SHOULD_BE_SHARED, 'ENV_SHOULD_BE_SHARED')
+    assert.equal(env.IS_OFFLINE, 'true')
   })
 })